functions/yt_ingest/requirements.txt
# /functions/yt_ingest/requirements.txt - Python dependencies for Cloud Function

# functions/yt_ingest/requirements.txt
functions-framework==3.8.2
Flask==3.1.0
google-cloud-storage==2.19.0
yt-dlp==2025.3.31
